# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Arokah API
  version: 1.0.0
  description: AI-powered personal stylist — daily looks, trip planning, cultural
    intelligence.
paths:
...
  /api/wardrobe/items/:
    get:
      operationId: wardrobe_items_list
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - wardrobe
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedClothingItemList'
          description: ''
    post:
      operationId: wardrobe_items_create
      tags:
      - wardrobe
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ClothingItemRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/ClothingItemRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClothingItem'
          description: ''
...
components:
  schemas:
...
    CategoryEnum:
      enum:
      - top
      - bottom
      - dress
      - outerwear
      - footwear
      - accessory
      - activewear
      - formal
      - other
      type: string
      description: |-
        * `top` - Top
        * `bottom` - Bottom
        * `dress` - Dress
        * `outerwear` - Outerwear
        * `footwear` - Footwear
        * `accessory` - Accessory
        * `activewear` - Activewear
        * `formal` - Formal
        * `other` - Other
    ClothingItem:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 200
        category:
          $ref: '#/components/schemas/CategoryEnum'
        formality:
          $ref: '#/components/schemas/FormalityEnum'
        season:
          $ref: '#/components/schemas/SeasonEnum'
        colors: {}
        material:
          type: string
          maxLength: 100
        weight_grams:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          nullable: true
        brand:
          type: string
          maxLength: 100
        image:
          type: string
          format: uri
          nullable: true
        image_clean:
          type: string
          format: uri
          nullable: true
        tags: {}
        is_active:
          type: boolean
        times_worn:
          type: integer
          readOnly: true
        last_worn:
          type: string
          format: date
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        user:
          type: integer
          readOnly: true
      required:
      - category
      - created_at
      - id
      - last_worn
      - name
      - times_worn
      - updated_at
      - user
    ClothingItemRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        category:
          $ref: '#/components/schemas/CategoryEnum'
        formality:
          $ref: '#/components/schemas/FormalityEnum'
        season:
          $ref: '#/components/schemas/SeasonEnum'
        colors: {}
        material:
          type: string
          maxLength: 100
        weight_grams:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          nullable: true
        brand:
          type: string
          maxLength: 100
        image:
          type: string
          format: binary
          nullable: true
        image_clean:
          type: string
          format: binary
          nullable: true
        tags: {}
        is_active:
          type: boolean
      required:
      - category
      - name
...
    FormalityEnum:
      enum:
      - casual
      - casual_smart
      - smart
      - formal
      - activewear
      type: string
      description: |-
        * `casual` - Casual
        * `casual_smart` - Casual Smart
        * `smart` - Smart
        * `formal` - Formal
        * `activewear` - Activewear
...
    PaginatedClothingItemList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/ClothingItem'
...
    SeasonEnum:
      enum:
      - spring
      - summer
      - autumn
      - winter
      - all
      type: string
      description: |-
        * `spring` - Spring
        * `summer` - Summer
        * `autumn` - Autumn
        * `winter` - Winter
        * `all` - All Season
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT